Rendering with an indices buffer successful

+++ app/src/main/java/com/arthurgichuhi/kotlinopengl/core/collada/dataStructures/MeshData.kt
@@ -4,7 +4,7 @@ data class MeshData(
     val vertices :FloatArray,
     val textureCords : FloatArray,
     var normals : FloatArray,
-    var indices : IntArray,
+    var indices : ShortArray,
     var jointIds : IntArray,
     var vertexWeights : FloatArray
 )
